Committer: Deniz Altinbuken <denizalti@rrdhcp60-69.redrover.cornell.edu>  Changes to be committed: 	modified:   Acceptor.py 	new file:   Client.py 	modified:   Commander.py 	modified:   Connection.py 	modified:   Group.py 	modified:   Leader.py 	new file:   Replica.py 	modified:   utils.py
The Replica is fully functioning: accepts performed operations and keeps state.
The Client is being added: The Client connects to a peer and makes either a debit or deposit request. If the debit/deposit request is accepted in paxos the transaction succeeds, otherwise it fails. Every node keeps the Bank state, which is the list of commands/transactions.
TODO:
-Client-
 The Client operation handling by an Acceptor: the requests are forwarded to a Leader.
 The Client operation handling by a Leader: a new paxos state is initiated, a reply is returned to the Client or forwarding Acceptor depending on the paxos outcome.
 State Keeping should be done better.

-Testing-
Paxos should be tested more thoroughly, corner test cases should be thought about, added.