Committer: Deniz Altinbuken <denizalti@rrdhcp60-388.redrover.cornell.edu>  On branch master  Changes to be committed: 	modified:   Acceptor.py 	new file:   Bank.py 	modified:   Client.py 	modified:   Connection.py 	modified:   Group.py 	modified:   Leader.py 	modified:   Peer.py 	modified:   Replica.py 	modified:   utils.py
Bank and Account Infrastructure fixed.
Operations spread along levels -

The Client connects with an account.
The Client makes a request to the Node.
The node forwards the request to the Leader.
Leader starts paxos.
When paxos succeeds Leader sends perform message.
Replica calls Bank for the command.
Bank calls the relevant account with the command.
Account makes the necessary changes.