All communication issues fixed, the types are changed accordingly, the Nodes work only with the port and for the bootstrap only the IP:Port is needed.
Paxos is working, the Replicas are the only ones to keep the Bank state. The Client interaction is fixed but the abstraction between the Paxos and the Bank is not complete yet.

An enum class is added to make the enumerations more python=like but it is not integrated with the code yet. (I wanted to show it to you first, looks a little redundant to me.)

File names decapitalized (learned that this is the pythonic way :))