Client is working now. node.py is slightly changed to support Client (Client keeps the server rather than the groups.)
Client is working, for now only the debit function is tested. The other commands will be similar.

There is a confusion about the Replica-Leader-Client communication. According to Robert's paper the Replica should be responding to the Client, which (in my understanding) would cause multiple responses. In the implementation (even if it's not working fully yet) we have the Leader talk to the Client depending on the responses it gets from replicas, hence two different messages:

MSG_RESPONSE: from Replica to Leader about the result of the command
MSG_CLIENTREPLY: from Leader to Client about the result of the clientrequest

All in all, there are still following major problems:

- Even if the Client is working, it is not fully tested (there are unimplemented functions).
- The response paradigm should be decided.
- The HELO, HELO_REPLY, NEW issue is not fixed, it is as we left before.
- The Commandnumber generation is dummy now, it is related to the communication between the Leader and Replica; that's why dummy for now. It will be fixed according to the choice about the responses.

P.S: I'm commiting code with a lot of XXX's, sorry about that.