Bug about keeping the socketset,nascentset fixed.
Bug: sockets are not being removed from the nascentset when a MSG_HELO was received, they are only removed when a timeout occurs. This was causing sockets to be added to the socketset multiple times (once for being used before and ones for being in the nascentset).
Fix: a socket is being added to the socketset only if it's not already in the set. the sockets are being removed from the nascentset only due to a timeout, they are still not being removed when a MSG_HELO is received. but it is not causing problems anymore.

Thus, multiple commands can be issued now.