The Response Collector keeps the list of replies and the sources and decides if the quorum is reached from the size of this list.
The Leader picks a pmax set (for all (commandnumber,proposal) pairs the ones with the highest ballotnumber. Initialized a PROPOSE STAGE for each.

The Response Collectors for the PROPOSE are indexed by the commandnumber now, as multiple PROPOSE messages can share a ballotnumber.

The Replica only executes the command next to execute and executes a command only if it was not executed before.