DNSServer and DNSQuery are created. They are not completed yet.
DNSServer binds to a UDP server socket and listens to DNSQueries.
A DNSQuery object can be created from the received data. The parsing
is being done according to rfc1035 section 4.

There is a test file: foo.py to test parsing.

A DNS Capture from Wireshark is also included, will be used for testing later.