Condition problem fixed.
Now when a client makes a notify the notified clients (with their commands) are added to the lock queue. This is done so that when they get an "UnusualReturn" it means they are awakened from a "wait" and they "can" acquire the Monitor Lock.

Replica is being edited to catch any type of exception that might be returned from the object-method. The exception will be passed along to the client and raised again in the ClientProxyObject. For this purpose I should add a bit to the ClientReply for the ProxyObject to check and raise a packed exception accordingly.