Zookeeper Objects created:
Bank Package
In this Package the classes related to the Bnak object can be found. The functionality
is as follows:
For every "Bank Object" a Persistant Node is created in the ZooKeeper server.
For every "Account" opened in the Bank a childnode is created under the banknode.
These childnodes keep serialized information about Accounts. Upon update the clients that
are connected to the same banknode are "notified" if they have the notifications on.
This way the related actions can be handled by the client.

This package also provides a template for similar objects that will be used to share and organize
data. Different Class (Object) definitions can be easily added, the Client and Shell
Threads can be revised to handle different operations.

Sync Package
In this Package the classes for synchronization objects can be found. There are Barrier, Lock,
Semaphore and Producer,Consumer Queue classes added.
The Sync client can be used to create and use any of these synchronization objects.

To give detail about the functionality the Barrier Object is designed as follows:
For every Barrier Object that is created a new barriernode is created under the syncnode.
This barriernode has the information about the barrier (size, name) serialized and saved in
the node. For every client that "enters" the barriernode an Ephemeral-Sequential childnode
is created. When the number of childnodes reach the size of the barrier the clients are notified.

Lock, Semaphore and Queue objects follow the same idea.