Partly implemented transitive invocations, there are circular import problems.
I have to clean the code and structure it in a better, cleaner way. As it got bigger it became very messy.
This is going to be done after the deadline.

The implementation for transitive invocations is unser srctransitive. I wasn't sure about the idea at all so I
didn't want to rape the original code. Once I finish it I'll clean the code properly.

Now I will write the text about transitive invocations.