Server side objectediter using AST.
It only adds **kwargs to functions defitions under the class that
will be used as our object. It doesn't ass **kwargs if it is already
present.

This technique is cleaner and more professional than the string read
and compare method used previously. So I'll edit the old proxygenerator
to use AST similarly.

The script expects to inputs: -m modulename -o objectname