OpenReplicaInitializer is initializing the system given the configuration.
- Eligiblenodes keeps nodes that have Python version > 2.7
- Logging mechanism changed. We have a Logger class now. It keeps log to
a file. The name of this file is used to retrieve the IP and the Port of
the running process. (All other ways failed).
- execute commands in plmanager take an additional parameter: wait.
if wait is True, the function wauts for the process to return. Otherwise
it returns without waiting. This was needed to start node processes.
- I'm not using the Coordinator anymore, but the related changes are not
done completely. For now when a bootstrap gets a join request it starts a
Paxos round. It may forward the request to the Leader in case
it is not the leader. For now the joining node will try another bootstrap
if the first try is not successful.
- We should start system responsible for openreplicacoordobj tomorrow. Only
after it is setup and connected to the openreplica nameservers we can fully
finish the deployment code. It has holes now.