Code is made faster but mostly bugs cleaned:
1) Leader handover was broken, now a second replica takes over without a problem.
2) The client can reconnect to any new replice without a problem.
3) connection lookup fixed after getting rid of peers.
4) connection add and removal fixed for select cases.

New bugs:
i) If a leader dies, and comes back after another replica takes over, a previously received clientrequest causes a KeyError while removing it from pendingcommands. This happens if only this message is received again exactly after the update is done.

ii) If a clientrequest is received before the leader took over completely, the connection to the client is not added properly and messages are not sent. It is not determined when this happens yet.

This commit includes debugging.